requirements.txt
dnspython=2.2.0
pyperclip=1.8.2
